Add support for React Native
