+++ examples/containers/App.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import CounterApp from './CounterApp';
 import TodoApp from './TodoApp';
-import { createRedux, Provider } from 'redux';
+import { createRedux } from 'redux';
+import { Provider } from 'redux/react';
 import * as stores from '../stores';
 
 const redux = createRedux(stores);
