+++ package.json
@@ -42,6 +42,7 @@
     "eslint-plugin-react": "^2.3.0",
     "expect": "^1.6.0",
     "mocha": "^2.2.5",
+    "react": "^0.13.0",
     "react-hot-loader": "^1.2.7",
     "rimraf": "^2.3.4",
     "webpack": "^1.9.6",
@@ -51,8 +52,7 @@
     "babel-runtime": "^5.4.7",
     "envify": "^3.4.0",
     "invariant": "^2.0.0",
-    "lodash": "^3.9.3",
-    "react": "^0.13.0"
+    "lodash": "^3.9.3"
   },
   "browserify": {
     "transform": [
