+++ react.js
@@ -1,0 +1,1 @@
+export * from './lib/react';
