+++ src/index.js
@@ -2,14 +2,6 @@
 export createRedux from './createRedux';
 export createDispatcher from './createDispatcher';
 
-// Wrapper components
-export Provider from './components/Provider';
-export Connector from './components/Connector';
-
-// Higher-order components (decorators)
-export provide from './components/provide';
-export connect from './components/connect';
-
 // Utilities
 export compose from './utils/composeMiddleware';
 export composeStores from './utils/composeStores';
